+++ js/templates/components/reportBtn.html
@@ -1,0 +1,3 @@
+<button class="iconBtnTn clrP clrBr tx2 <% if(ob.reported) print('reported') %> js-reportBtn">
+  <i class="ion-ios-flag <% if(ob.reported) print('clrTErr') %>"></i>
+</button>
