--- js/templates/reportBtn.html
@@ -1,3 +1,0 @@
-<button class="iconBtnTn clrP clrBr tx2 js-reportBtn">
-  <i class="ion-ios-flag <% if(ob.reported) print('clrTErr') %>"></i>
-</button>
